Move build-plan packages
