+++ jib-build-plan/src/main/java/com/google/cloud/tools/jib/api/buildplan/FileEntry.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.buildplan;
+package com.google.cloud.tools.jib.api.buildplan;
 
 import java.nio.file.Path;
 import java.time.Instant;
