+++ jib-build-plan/src/main/java/com/google/cloud/tools/jib/api/buildplan/FilePermissions.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.buildplan;
+package com.google.cloud.tools.jib.api.buildplan;
 
 import java.nio.file.attribute.PosixFilePermission;
 import java.util.Collections;
